Add Gitmux link
First, thanks for doing that awesome list!
I'm the author and maintainer of https://github.com/arl/gitmux. Gitmux shows Git status in tmux status bar, it's a rewrite from scratch of https://github.com/arl/tmux-gitbar, which I also wrote. I rewrote it for various reasons, main ones being that it was written in bash script and it was becoming  increasingly difficult to maintain and add the features, plus it was only working with bash shells.
Gitmux however has already all the features it needs and is shell independent. 

Gitmux is shell-indepedent and some binaries for many platforms are pre-compiled and available as Github releases.
It's very stable and used everyday by many persons already, it has few stars since I published it like recently.
I believe gitmux is a good match with this list ;-)
I know the project only has 24 stars for now but tmux-gitbar (which is a rewrite of) had 160.